requirements.txt
pytest
matplotlib
Pillow